remove schema from compartment and from all composites. schema is now generated by initialize_state, and saved in Store
