+++ vivarium/composites/variable_flagella.py
@@ -5,7 +5,7 @@ import os
 import random
 
 from vivarium.compartment.process import initialize_state
-from vivarium.compartment.composition import get_derivers, get_schema
+from vivarium.compartment.composition import get_derivers
 
 # processes
 from vivarium.processes.Endres2006_chemoreceptor import ReceptorCluster
@@ -70,17 +70,13 @@ def compose_variable_flagella(config):
     processes.extend(derivers['deriver_processes'])  # add deriver processes
     topology.update(derivers['deriver_topology'])  # add deriver topology
 
-    # get schema
-    schema = get_schema(processes, topology)
-
     # initialize the states
-    states = initialize_state(processes, topology, schema, config.get('initial_state', {}))
+    states = initialize_state(processes, topology, config.get('initial_state', {}))
 
     options = {
         'environment_port': 'environment',
         # 'exchange_port': 'exchange',
         'topology': topology,
-        'schema': schema,
         'initial_time': config.get('initial_time', 0.0),
         'divide_condition': divide_condition}
